testing/web-platform/tests/shadow-dom/elements-and-dom-objects/extensions-to-event-interface/event-path-002.html

<!DOCTYPE html>









<html>
<head>
<title>Shadow DOM Test - event path</title>
<link rel="author" title="Kazuhito Hokamura" href="mailto:k.hokamura@gmail.com">
<link rel="help" href="https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/shadow/index.html#extensions-to-event">
<meta name="assert" content="Extensions to Event Interface: event.path is readonly">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
var t = async_test('event.path is readonly');

t.step(unit(function(ctx) {
    var doc = newRenderedHTMLDocument(ctx);
    var div = doc.createElement('div');
    doc.body.appendChild(div);

    div.addEventListener('click', t.step_func(function(e) {
        var obj = {};
        e.path = obj;
        assert_not_equals(e.path, obj);

        t.done();
    }));

    var event = doc.createEvent('HTMLEvents');
    event.initEvent('click', true, false);
    div.dispatchEvent(event);
}));
</script>
</body>
</html>
